chore: remove devtools from production
envify build script by passing in variables using the DefinePlugin
(details:
https://github.com/webpack/docs/wiki/list-of-plugins#defineplugin)
