+++ src/configureStore.prod.js
@@ -1,0 +1,6 @@
+import {combineReducers, createStore} from 'redux'
+import * as reducers from './reducers'
+
+export default function configureStore () {
+  return createStore(combineReducers(reducers))
+}
